+++ packages/cdktf-cli/lib/get/generator/emitter/resource-emitter.ts
@@ -13,7 +13,7 @@ export class ResourceEmitter {
     this.code.line();
     this.code.line(`// Resource`)
     this.code.line();
-    this.code.openBlock(`export class ${resource.className} extends ${resource.parentClassName}`);
+    this.code.openBlock(`export class ${resource.className} extends cdktf.${resource.parentClassName}`);
 
     this.emitHeader('INITIALIZER');
     this.emitInitializer(resource);
@@ -41,7 +41,7 @@ export class ResourceEmitter {
     this.code.open(`return {`);
 
     for (const att of resource.synthesizableAttributes) {
-      this.code.line(`${att.terraformName}: this.${att.storageName},`);
+      this.attributesEmitter.emitToTerraform(att, false);
     }
 
     this.code.close(`};`);
@@ -50,7 +50,9 @@ export class ResourceEmitter {
 
   private emitResourceAttributes(resource: ResourceModel) {
     for (const att of resource.attributes) {
-      this.attributesEmitter.emit(att)
+      this.attributesEmitter.emit(att, 
+        this.attributesEmitter.needsResetEscape(att, resource.attributes), 
+        this.attributesEmitter.needsInputEscape(att, resource.attributes));
     }
   }