+++ packages/cdktf/test/data-source.test.ts
@@ -1,4 +1,4 @@
-import { TerraformStack, Testing } from '../lib';
+import { TerraformStack, Testing, Token } from '../lib';
 import { TestResource } from './helper';
 import { TestDataSource } from './helper/data-source';
 
@@ -26,3 +26,60 @@ test('with complex computed list', () => {
   expect(Testing.synth(stack)).toMatchSnapshot();
 });
 
+test('with string map', () => {
+  const app = Testing.app();
+  const stack = new TerraformStack(app, 'test-data-source');
+
+  const dataSource = new TestDataSource(stack, 'test', {
+    name: 'foo'
+  });
+  new TestResource(stack, 'test-resource', {
+    name: dataSource.stringMap('id'),
+  })
+
+  expect(Testing.synth(stack)).toMatchSnapshot();
+});
+
+test('with number map', () => {
+  const app = Testing.app();
+  const stack = new TerraformStack(app, 'test-data-source');
+
+  const dataSource = new TestDataSource(stack, 'test', {
+    name: 'foo'
+  });
+  new TestResource(stack, 'test-resource', {
+    name: Token.asString(dataSource.numberMap('id')),
+  })
+
+  expect(Testing.synth(stack)).toMatchSnapshot();
+});
+
+test('with boolean map', () => {
+  const app = Testing.app();
+  const stack = new TerraformStack(app, 'test-data-source');
+
+  const dataSource = new TestDataSource(stack, 'test', {
+    name: 'foo'
+  });
+  new TestResource(stack, 'test-resource', {
+    name: dataSource.booleanMap('id').toString(),
+  })
+
+  expect(Testing.synth(stack)).toMatchSnapshot();
+});
+
+test('dependent data source', () => {
+  const app = Testing.app();
+  const stack = new TerraformStack(app, 'test');
+
+  const resource = new TestResource(stack, 'resource', {
+    name: 'foo'
+  })
+
+  new TestDataSource(stack, 'data_source', {
+    name: 'foo',
+    dependsOn: [resource]
+  })
+
+  expect(Testing.synth(stack)).toMatchSnapshot();
+});