+++ CHANGELOG.md
@@ -1,3 +1,32 @@
+## 0.0.19
+
+**Fixed bugs:**
+
+- Depends on expansion [\#442](https://github.com/hashicorp/terraform-cdk/issues/442)
+- Make sure tokens resolved before converting to json [\#439](https://github.com/hashicorp/terraform-cdk/issues/439)
+- Alter reset and input names when they conflict [\#429](https://github.com/hashicorp/terraform-cdk/issues/429)
+- Fix complex map attribute data source reference [\#441](https://github.com/hashicorp/terraform-cdk/issues/441)
+- Remap 'equals' to 'equalTo' to work around jsii limitation [\#428](https://github.com/hashicorp/terraform-cdk/issues/428)
+- Generated custom mappings [\#395](https://github.com/hashicorp/terraform-cdk/issues/395)
+
+## 0.0.18
+
+**Fixed bugs:**
+
+- Update jsii with python docs fix [\#418](https://github.com/hashicorp/terraform-cdk/issues/418)
+- Maintain dependencies between required construct properties [\#310](https://github.com/hashicorp/terraform-cdk/issues/310)
+
+**Implemented enhancements:**
+
+- Java support [\#360](https://github.com/hashicorp/terraform-cdk/issues/360)
+- Simple Terraform Locals [\#357](https://github.com/hashicorp/terraform-cdk/issues/357)
+- allow to use TERRAFORM_BINARY_NAME env to set terraform binary [\#366](https://github.com/hashicorp/terraform-cdk/issues/366)
+- Default to NPM in templates [\#359](https://github.com/hashicorp/terraform-cdk/issues/359)
+- Terraform version check [\#378](https://github.com/hashicorp/terraform-cdk/issues/378)
+- Terraform Variables [\#277](https://github.com/hashicorp/terraform-cdk/issues/277)
+- Terraform hcl module [\#375](https://github.com/hashicorp/terraform-cdk/issues/375)
+- Support for matching existing Terraform names [\#329](https://github.com/hashicorp/terraform-cdk/issues/329)
+
 ## 0.0.17
 
 **Fixed bugs:**